BAU: Fix deployment issues caused by log group conflicts
The lambda resources need to depend on the corresponding log group resources to ensure that the lambda service doesn't attempt to create a duplicate log group and cause an erorr during deployment. Creating the dependency ensures that the log group gets created before the lambda and there is no conflict.