BAU: Minor refactor to BearerTokenRetrievalStateMachineAlarm
The alarm transitions to OK every morning due to MissingData being treated as missing. The alarm triggers a slack messages causing the channel to get spammed.
Refactored to treatMissingData as notBreaching.

Also, previously the alarm did not match the description, rather than 4 failures within an hour, it was atleast 1 failure in 4 different 5 minute periods within an hour. This would mean if for example, if 100 executions failed but only within a 5 minute period it would not trigger the alarm.
Refactored to make the alarm match the description.